lambda-backend/website-pf/requirements.txt
Flask==1.1.2
boto3==1.14.56
mysql-connector-python==8.0.21